Add files via upload
We all know that JavaScript makes the Website alive, It makes it feel like working or doing some activity. 
Here In the dice game, its first to let the computer choose a number between 1-6 (a dice has 6 sides), then we name the images as such as they are easy to fetch  by the computer logically. We then connect the choosen number with the png of the image. Now sorting dice 1, the last thing left was to link them together which was helped by the set attribute of JavaScript , And here we are done.
This Mini Project took me around an hour since I really needed to understand the knowledge of how and what the codes really are doing.  
I even complied my doubts to check the errors which occurred which seemed most important which let me took 1 whole hour to complete it. 

I hope you liked my first attempt of Javascript.
Will come with more minor projects. 
Thankyou :)